work on spawns / workers
